Slim down surface of loader system
